requirements.txt
scrapy==2.8.0
langchain==0.0.267
openai==1.3.5
faiss-cpu==1.7.4
# Use a newer PyMuPDF version that's more compatible with Python 3.13
pymupdf
numpy
pandas
scikit-learn
rank-bm25==0.2.2
streamlit==1.28.0
python-dotenv==1.0.0
arxiv==1.4.8
scholarly==1.7.11
beautifulsoup4==4.12.2
tiktoken==0.5.1
